requirements.txt
# 统一的依赖文件 - Smart Home Voice Assistant

# --- 核心依赖 ---
# FastAPI后端
fastapi==0.115.12
uvicorn[standard]==0.34.2
python-multipart==0.0.6
pydantic==2.11.4
pyyaml==6.0.2
requests==2.30.0

# --- NLP服务依赖 ---
# STT依赖
numpy>=1.21.6,<1.28.0
torch==2.3.1
torchaudio==2.3.1
transformers==4.51.3
openai-whisper==20240930
soundfile==0.12.1
PyAudio==0.2.14
dataoceanai-dolphin==20250507  # Dolphin STT引擎依赖

# NLU依赖
scikit-learn==1.6.1
jieba==0.42.1
langdetect==1.0.9

# TTS依赖
edge-tts==7.0.2
pyttsx3==2.98

# --- 工具和实用程序 ---
python-dotenv==1.0.0
websockets==11.0.3
cffi==1.16.0
setuptools-rust==1.11.1 